+++ realm-transformer/src/main/groovy/io/realm/transformer/BytecodeModifier.groovy
@@ -66,11 +66,10 @@ class BytecodeModifier {
                     !behavior.name.startsWith('realmGet$') &&
                     !behavior.name.startsWith('realmSet$')
                 ) || (
-                    behavior instanceof CtConstructor &&
-                    !modelClasses.contains(clazz)
+                    behavior instanceof CtConstructor
                 )
             ) {
-                behavior.instrument(new FieldAccessToAccessorConverter(managedFields, clazz, behavior))
+                behavior.instrument(new FieldAccessToAccessorConverter(managedFields, clazz, behavior, modelClasses.contains(clazz)))
             }
         }
     }
@@ -94,11 +93,18 @@ class BytecodeModifier {
         final List<CtField> managedFields
         final CtClass ctClass
         final CtBehavior behavior
+        final boolean isModelClass
+        final boolean isInConstructor
 
-        FieldAccessToAccessorConverter(List<CtField> managedFields, CtClass ctClass, CtBehavior behavior) {
+        FieldAccessToAccessorConverter(List<CtField> managedFields,
+                                       CtClass ctClass,
+                                       CtBehavior behavior,
+                                       boolean isModelClass) {
             this.managedFields = managedFields
             this.ctClass = ctClass
             this.behavior = behavior
+            this.isModelClass = isModelClass
+            this.isInConstructor = behavior instanceof CtConstructor
         }
 
         @Override
@@ -112,9 +118,19 @@ class BytecodeModifier {
                 logger.info "        Methods: ${ctClass.declaredMethods}"
                 def fieldName = fieldAccess.fieldName
                 if (fieldAccess.isReader()) {
-                    fieldAccess.replace('$_ = $0.realmGet$' + fieldName + '();')
+                    if (isInConstructor && isModelClass) {
+                        // work around https://github.com/realm/realm-java/issues/2536
+                        fieldAccess.replace('$_ = ($0 == this) ? $0.' + fieldName + ' : $0.realmGet$' + fieldName + '();')
+                    } else {
+                        fieldAccess.replace('$_ = $0.realmGet$' + fieldName + '();')
+                    }
                 } else if (fieldAccess.isWriter()) {
-                    fieldAccess.replace('$0.realmSet$' + fieldName + '($1);')
+                    if (isInConstructor && isModelClass) {
+                        // work around https://github.com/realm/realm-java/issues/2536
+                        fieldAccess.replace('if ($0 == this) {$0.' + fieldName + ' = $1;} else { $0.realmSet$' + fieldName + '($1);}')
+                    } else {
+                        fieldAccess.replace('$0.realmSet$' + fieldName + '($1);')
+                    }
                 }
             }
         }